rename file and update tests
